requirements.txt
langchain
langchain_community
nomic
pypdf2
langchain_core
langchain_nomic
streamlit
google
langchain_google_genai
python-dotenv
faiss-cpu
transformers
accelerate
bitsandbytes
unsloth
datasets
trl
peft
fastapi
pydantic
typing